Improve SQLite API and add indexes
